requirements.txt
certifi==2023.7.22
influxdb-client==1.37.0
numpy==1.25.2
pandas==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
reactivex==4.0.4
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
python-dotenv==1.0.0
